added google login

+++ vagrant/catalog/finalproject.py
@@ -2,11 +2,174 @@ import controller
 
 from datetime import datetime, date
 from database import Department, Base, Seminar
-from flask import Flask, render_template, request, redirect, url_for, flash, jsonify
+from flask import Flask, render_template, request, redirect, url_for, \
+                  flash, jsonify, make_response
+# These import are for anti-forgery state token creation 
+from flask import session as login_session
+import random, string
+# These imports are for Gconnect
+from oauth2client.client import flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+import httplib2
+import json
+import requests
+
+CLIENT_ID = json.loads(
+    open('client_secrets.json', 'r').read())['web']['client_id']
+APPLICATION_NAME = "Seminar App"
+
 # create instance of class with name of running application as arg
 # anytime we run an application in python a special variable called
 # name gets defined for the application an all of imports it uses
 app = Flask(__name__)
+
+
+# Create anti-forgery state token
+@app.route('/login')
+def showLogin():
+    state = ''.join(random.choice(string.ascii_uppercase + string.digits)
+                    for x in xrange(32))
+    login_session['state'] = state
+    # return "The current session state is %s" % login_session['state']
+    return render_template('login.html', STATE=state)
+
+
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorization code
+    code = request.data
+
+    try:
+        # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1])
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print "Token's client ID does not match app's."
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_access_token = login_session.get('access_token')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_access_token is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already connected.'),
+                                 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['access_token'] = credentials.access_token
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = answer.json()
+
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+    login_session['email'] = data['email']
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    flash("you are now logged in as %s" % login_session['username'])
+    print "done!"
+    return output
+
+# DISCONNECT - Revoke a current user's token and reset their login_session
+@app.route('/gdisconnect')
+def gdisconnect():
+    access_token = login_session['access_token']
+    print 'In gdisconnect access token is %s', access_token
+    print 'User name is: ' 
+    print login_session['username']
+    if access_token is None:
+ 	print 'Access Token is None'
+    	response = make_response(json.dumps('Current user not connected.'), 401)
+    	response.headers['Content-Type'] = 'application/json'
+    	return response
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    print 'result is '
+    print result
+    if result['status'] == '200':
+	del login_session['access_token'] 
+    	del login_session['gplus_id']
+    	del login_session['username']
+    	del login_session['email']
+    	del login_session['picture']
+    	response = make_response(json.dumps('Successfully disconnected.'), 200)
+    	response.headers['Content-Type'] = 'application/json'
+    	return response
+    else:
+	
+    	response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+    	response.headers['Content-Type'] = 'application/json'
+    	return response
+
+
+# User Helper Functions
+
+def createUser(login_session):
+    newUser = User(name=login_session['username'], email=login_session[
+                   'email'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
 
 # decorator wraps function into app.route function if any of these
 # addresses get entered, the HelloWorld function gets executed
@@ -29,6 +192,9 @@ def departmentSeminars(department_id):
 
 @app.route('/Deparment/new/', methods=['GET','POST'])
 def newDepartment():
+    # Check to see if user is logged in to get to page
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method =='POST':
         if request.form['name'] and request.form['description']:
             editName = request.form['name']
@@ -44,6 +210,9 @@ def newDepartment():
 
 @app.route('/Deparment/<int:department_id>/edit/', methods=['GET','POST'])
 def editDepartment(department_id):
+    # Check to see if user is logged in to get to page
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method =='POST':
         if request.form['name']:
             editName = request.form['name']
@@ -62,6 +231,9 @@ def editDepartment(department_id):
 
 @app.route('/Deparment/<int:department_id>/delete/', methods=['GET','POST'])
 def deleteDepartment(department_id):
+    # Check to see if user is logged in to get to page
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method =='POST':
         # Delete department to controller
         controller.deleteDepartment(department_id)
@@ -74,6 +246,9 @@ def deleteDepartment(department_id):
 
 @app.route('/Deparment/<int:department_id>/new/', methods=['GET','POST'])
 def newSeminarItem(department_id):
+    # Check to see if user is logged in to get to page
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
         if request.form['title'] and request.form['date_time']:
             dept = controller.getDepartment(department_id)
@@ -81,7 +256,7 @@ def newSeminarItem(department_id):
             # Form fields
             addTitle = request.form['title']
             addSpeaker = request.form['speaker']
-            addDate = datetime.strptime(request.form['date_time'], '%m/%d/%Y').date()
+            addDate = datetime.strptime(request.form['date_time'], '%d, %B %Y').date()
             addAbstract = request.form['abstract']
             addBuilding = request.form['building']
             addRoom = request.form['room']
@@ -115,13 +290,16 @@ def newSeminarItem(department_id):
 
 @app.route('/Deparment/<int:department_id>/<int:seminar_id>/edit/', methods=['GET','POST'])
 def editSeminarItem(department_id, seminar_id):
+    # Check to see if user is logged in to get to page
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method =='POST':
               
         dept = controller.getDepartment(department_id)
         # Form fields
         addTitle = request.form['title']
         addSpeaker = request.form['speaker']
-        addDate = datetime.strptime(request.form['date_time'], '%Y-%m-%d').date()
+        addDate = datetime.strptime(request.form['date_time'], '%d, %B %Y').date()
         addAbstract = request.form['abstract']
         addBuilding = request.form['building']
         addRoom = request.form['room']
@@ -151,8 +329,18 @@ def editSeminarItem(department_id, seminar_id):
 
 # Task 3: Create a route for deleteSeminarItem function here
 
-@app.route('/Deparment/<int:department_id>/<int:seminar_id>/delete/', methods=['GET','POST'])
+@app.route('/Deparment/<int:department_id>/<int:seminar_id>/delete/', 
+            methods=['GET','POST'])
 def deleteSeminarItem(department_id, seminar_id):
+    getSeminarUser = ""
+    # Check to see if user is logged in to get to page
+    if 'username' not in login_session:
+        return redirect('/login')
+    if getSeminarUser.user_id != login_session['user_id']:
+        return """<script>function myFunction() { alert('You are not authorized
+                  to delete this Seminar. Please create your own seminar in 
+                  order to delete.');}</script><body onload='myFunction()'>
+               """
     if request.method =='POST':
         # Delete department to controller
         controller.deleteSeminar(seminar_id)
@@ -167,11 +355,16 @@ def deleteSeminarItem(department_id, seminar_id):
                                 department=getdepartment,
                                 seminar=getseminar)
 
-# Making an API Endpoint (GET Request)
+# Making an API Endpoint (GET Request) for JSON
 @app.route('/Department/<int:department_id>/seminar/JSON')
 def departmentSeminarJSON(department_id):
     listOfSeminars = controller.getAllSeminarItems(department_id)   
     return jsonify(Seminars=[i.serialize for i in listOfSeminars])
+
+@app.route('/Department/JSON')
+def departmentJSON():
+    listOfDeparments = controller.getAllDepartments()
+    return jsonify(Departments=[i.serialize for i in listOfDeparments])
 
 
 # the application run by the python interpreter gets a name variable
